requirements.txt
robotframework==6.1.1
robotframework-browser==17.3.0
robotframework-pabot==2.16.0
robotframework-requests==0.9.5
robotframework-seleniumlibrary==6.1.2
robotframework-tidy==4.4.0
webdriver-manager==4.0.0
robotframework-lsp==1.11.0
selenium==4.12.0
black==23.9.1
robotframework-tidy==4.4.0